ai_assistant/py/requirements.txt
flask
flask-cors
ollama
mistralai
openai
anthropic
pyOpenSSL
google-generativeai
python-weather
pycouchdb
pyttsx3
faster-whisper
pydub
ffmpeg